feat(Goal): Add deleted verification when delete

+++ src/main/java/ac/dnd/dodal/domain/goal/model/Goal.java
@@ -57,6 +57,9 @@ public class Goal extends BaseEntity {
         if (this.userId != userId) {
             throw new UnauthorizedException();
         }
+        if (this.deletedAt != null) {
+            throw new BadRequestException(GoalExceptionCode.GOAL_ALREADY_DELETED);
+        }
         super.delete();
     }
